% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sentopics-conversions}
\alias{sentopics-conversions}
\alias{as.sentopicmodel}
\alias{as.LDA}
\alias{as.rJST}
\alias{as.JST}
\title{Title}
\usage{
as.sentopicmodel(x)

as.LDA(x)

as.rJST(x)

as.JST(x)
}
\arguments{
\item{x}{...}
}
\value{
a relevant \code{sentopicmodel} object
}
\description{
Title
}
