% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{compute_PicaultRenaut_scores}
\alias{compute_PicaultRenaut_scores}
\title{Compute scores using the Picault-Renault lexicon}
\usage{
compute_PicaultRenaut_scores(
  x,
  min_ngram = 2,
  by_doc = TRUE,
  return_dfm = FALSE
)
}
\arguments{
\item{x}{...}

\item{min_ngram}{...}

\item{by_doc}{...}

\item{return_dfm}{...}
}
\description{
...
}
