% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{ECB_press_conferences}
\alias{ECB_press_conferences}
\title{Corpus of press conferences from the European Central Bank}
\format{
A \link[quanteda:corpus]{quanteda::corpus} object.
}
\source{
\url{https://www.ecb.europa.eu/press/pubbydate/html/index.en.html}.
}
\usage{
ECB_press_conferences
}
\description{
A corpus of 260 ECB press conference, split into 4224
paragraphs. The corpus contains a \emph{docvars} indicating the date
of the press conference.
}
\examples{
docvars(ECB_press_conferences)

}
\seealso{
\link{ECB_press_conferences_tokens}
}
\keyword{datasets}
