% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{ECB_press_conferences_tokens}
\alias{ECB_press_conferences_tokens}
\title{Tokenized press conferences}
\format{
A \link[quanteda:tokens]{quanteda::tokens} object.
}
\source{
\url{https://www.ecb.europa.eu/press/pubbydate/html/index.en.html?name_of_publication=Monetary\%20policy\%20statement}.
}
\usage{
ECB_press_conferences_tokens
}
\description{
The pre-processed and tokenized version of the
\link{ECB_press_conferences} corpus of press conferences. The processing
involved the following steps:
\itemize{
\item Subset paragraphs shorter than 10 words
\item Removal of stop words
\item Part-of-speech tagging, following which only nouns, proper nouns and
adjective were retained.
\item Detection and merging of frequent compound words
\item Frequency-based cleaning of rare and very common words
}
}
\examples{
LDA(ECB_press_conferences_tokens)

}
\seealso{
\link{ECB_press_conferences}
}
\keyword{datasets}
