% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{ECB_speeches}
\alias{ECB_speeches}
\title{Example tokenized speeches}
\format{
A \code{quanteda::tokens} object.
}
\source{
\url{www.ecb.europa.eu/press/key/html/index.en.html}.
}
\usage{
data("ECB_speeches")
}
\description{
...
}
\examples{
data("ECB_speeches")
head(ECB_speeches)

}
\keyword{datasets}
