% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_ECB_press_conferences}
\alias{get_ECB_press_conferences}
\title{Download press conferences from the European Central Bank}
\usage{
get_ECB_press_conferences(
  years = 1998:2021,
  language = "en",
  data.table = TRUE
)
}
\arguments{
\item{years}{the years for which press conferences should be retrieved}

\item{language}{the language in which press conferences should be retrieved}

\item{data.table}{if TRUE, returns a \link{data.table}. Otherwise, return a list
in which each element is a press conference.}
}
\value{
Depending on the arguments, returns either a data.frame or a
\link[quanteda:tokens]{quanteda::tokens} object containing press conferences of the ECB.
}
\description{
This helper function automatically retrieve the full data set of
press conferences made available by the ECB. It implements a number of
pre-processing steps used to remove the Q&A section from the text.
}
