% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{LSDfr}
\alias{LSDfr}
\title{Example lexicon2}
\format{
A quanteda \code{dictionary}.
}
\source{
\url{https://www.poltext.org/fr/donnees-et-analyses/lexicoder}
}
\usage{
data("LSDfr")
}
\description{
...
}
\examples{
data("LSDfr")
summary(LSDfr)

}
\keyword{datasets}
