% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{PicaultRenault}
\alias{PicaultRenault}
\title{Picault-Renault lexicon}
\format{
A \code{data.table}.
}
\source{
\url{http://www.cbcomindex.com/lexicon.php}
}
\usage{
data("PicaultRenault")
}
\description{
...
}
\examples{
data("PicaultRenault")
summary(PicaultRenault)

}
\keyword{datasets}
