% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sentopicmodel}
\alias{print.sentopicmodel}
\alias{print.rJST}
\alias{print.LDA}
\alias{print.JST}
\title{Print method for sentopics models}
\usage{
\method{print}{sentopicmodel}(x, extended = FALSE, ...)

\method{print}{rJST}(x, extended = FALSE, ...)

\method{print}{LDA}(x, extended = FALSE, ...)

\method{print}{JST}(x, extended = FALSE, ...)
}
\arguments{
\item{x}{the model to be printed}

\item{extended}{if \code{TRUE}, extends the print to include some helpful related
functions. Automatically displayed once per session.}

\item{...}{not used}
}
\description{
Print methods for \strong{sentopics} models. Once per session (or
forced by using \code{extended = TRUE}), it lists the most important function
related to \strong{sentopics} models.
}
