% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{proportion_topics}
\alias{proportion_topics}
\alias{plot_proportion_topics}
\title{Title}
\usage{
proportion_topics(
  x,
  period = c("year", "month", "day"),
  rolling_window = 1,
  plot = c(FALSE, TRUE, "silent"),
  plot_ridgelines = TRUE,
  as.xts = TRUE,
  ...
)

plot_proportion_topics(
  x,
  period = c("year", "month", "day"),
  rolling_window = 1,
  plot_ridgelines = TRUE,
  ...
)
}
\arguments{
\item{x}{...}

\item{period}{...}

\item{rolling_window}{...}

\item{plot}{...}

\item{plot_ridgelines}{...}

\item{as.xts}{...}

\item{...}{...}
}
\value{
...
}
\description{
Title
}
\seealso{
Other plot functions: 
\code{\link{sentiment_breakdown}()},
\code{\link{sentiment_topics}()}
}
\concept{plot functions}
