% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R, R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.tokens}
\alias{melt}
\alias{docvars}
\alias{tokens}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:melt.data.table]{melt}}}

  \item{quanteda}{\code{\link[quanteda]{as.tokens}}, \code{\link[quanteda]{docvars}}, \code{\link[quanteda]{tokens}}}
}}

