% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sentopicmodel}
\alias{sentopicmodel}
\title{Create a sentopic model}
\usage{
sentopicmodel(
  x,
  lexicon = NULL,
  L1 = 5,
  L2 = 3,
  L1prior = 1,
  L2prior = 5,
  beta = 0.01,
  L1cycle = 0,
  L2cycle = 0,
  initLDA = 0,
  smooth = 0,
  reversed = TRUE
)
}
\arguments{
\item{x}{tokens object containing the texts. A coercion will be attempted if \code{x} is not a tokens.}

\item{lexicon}{a \code{quanteda} dictionary with positive and negative categories}

\item{L1}{the number of labels in the first document mixture layer}

\item{L2}{the number of labels in the second document mixture layer}

\item{L1prior}{the first layer hyperparameter of document mixtures}

\item{L2prior}{the second layer hyperparameter of document mixtures}

\item{beta}{the hyperparameter of vocabulary distribution}

\item{L1cycle}{integer specifying the cycle size between two updates of the hyperparameter L1prior}

\item{L2cycle}{integer specifying the cycle size between two updates of the hyperparameter L2prior}

\item{initLDA}{integer specifying the number of iterations of the LDA burn-in}

\item{smooth}{integer specyfing the number of iterations of the smoothed burn-in}
}
\value{
An initialized \code{sentopicmodel} object inheriting the relevant model class.
This object corresponds to a Gibbs sampler estimation with zero iterations.
The MCMC can be iterated using the \code{\link[=grow]{grow()}} function.

The estimated distributions of the model are accessible using the dollar-sign operator.
The \code{\link[=topWords]{topWords()}} function easily extract the most probables words of each topic/sentiment.
}
\description{
The set of functions \code{\link[=LDA]{LDA()}}, \code{\link[=JST]{JST()}}, \code{\link[=rJST]{rJST()}} and \code{\link[=sentopicmodel]{sentopicmodel()}} are all wrappers to an unified C++
routine and attempt to replicate their corresponding model.
}
\examples{
data(ECB_speeches)
LDA(ECB_speeches)
rJST(ECB_speeches, lex = LoughranMcDonald)
}
\seealso{
\code{\link[=grow]{grow()}}, \code{\link[=topWords]{topWords()}}

Other topic models: 
\code{\link{JST}()},
\code{\link{LDA}()},
\code{\link{rJST}()}
}
\author{
Olivier Delmarcelle
}
\concept{topic models}
\keyword{internal}
