% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sentopics-conversions}
\alias{sentopics-conversions}
\alias{as.sentopicmodel}
\alias{as.rJST}
\alias{as.JST}
\title{Internal conversions between \strong{sentopics} models.}
\usage{
as.sentopicmodel(x)

as.rJST(x)

as.JST(x)
}
\arguments{
\item{x}{A \strong{sentopics} model}
}
\value{
A \strong{sentopics} model of the relevant class
}
\description{
Internal conversions between \strong{sentopics} models.
}
\keyword{internal}
