% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{package}
\name{sentopics-package}
\alias{sentopics}
\alias{sentopics-package}
\title{sentopics: tools for estimating and analyzing various classes of sentiment/topic models}
\description{
...
}
\note{
Please cite the package in publications. Use \code{citation("sentopics")}.
}
\references{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odelmarcelle/sentopics}
  \item Report bugs at \url{https://github.com/odelmarcelle/sentopics/issues}
}

}
\author{
\strong{Maintainer}: Olivier Delmarcelle \email{delmarcelle.olivier@gmail.com}

Authors:
\itemize{
  \item Samuel Borms \email{samuel.borms@unine.ch} (\href{https://orcid.org/0000-0001-9533-1870}{ORCID})
}

}
