% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{sentopics_labels}
\alias{sentopics_labels}
\title{Title}
\usage{
sentopics_labels(x)
}
\arguments{
\item{x}{a \code{sentopicmodel} created from the \code{\link[=LDA]{LDA()}}, \code{\link[=JST]{JST()}}, \code{\link[=rJST]{rJST()}} or
\code{\link[=sentopicmodel]{sentopicmodel()}} function}
}
\description{
Compute, extract or replace the internal sentiment of a
\code{sentopicmodel}. The internal sentiment is used to create time series using
the functions \code{\link[=sentiment_series]{sentiment_series()}} or \code{\link[=sentiment_topics]{sentiment_topics()}}. If the input
model contains a sentiment layer, sentiment can be computed directly from
the output of the model. Otherwise, sentiment obtained externally should be
added for each document.
}
\examples{
jst <- JST(ECB_speeches, lexicon = LoughranMcDonald)
sentopics_sentiment(jst)

jst <- grow(jst, 10)
sentopics_sentiment(jst, override = TRUE)

## Multivariate sentiment for rJST models
rjst <- rJST(ECB_speeches, lexicon = LoughranMcDonald)
rjst <- grow(rjst, 10)
sentopics_sentiment(rjst, override = TRUE)







}
\seealso{
Other sentopics helpers: 
\code{\link{sentopics_date}()},
\code{\link{sentopics_sentiment}()}
}
\author{
Olivier Delmarcelle
}
\concept{sentopics helpers}
