% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{vocabFromList}
\alias{vocabFromList}
\title{Generate vocabulary from list of words}
\usage{
vocabFromList(list)
}
\arguments{
\item{list}{list of words. First list level determine the topic. Second list
level indicate the sentiment}
}
\value{
an array usable in \code{generateDocuments}
}
\description{
Generate vocabulary from list of words
}
\examples{
sentopics:::vocabFromList(list(
  list(c("tasty", "delicious"), c("bad", "smelly")),
  list(c("surprising", "amazing"), c("boring", "annoying")),
  list("fearless", "coward")
))
sentopics:::vocabFromList(lapply(1:3, function(i) c(paste0("topic",i,"w1"))))
}
\keyword{internal}
