% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.LDA}
\alias{as.LDA}
\alias{as.LDA.STM}
\title{Conversions from other packages to LDA}
\usage{
as.LDA(x, ...)

\method{as.LDA}{STM}(x, docs, ...)
}
\arguments{
\item{x}{an estimated topic model from the \pkg{stm} package.}

\item{docs}{for the conversion from \pkg{stm}, the documents used to create
the object.}
}
\description{
These functions converts estimated models from other topic
modeling packages to the format used by sentopics.
}
\details{
The models initially estimated through variational inference suffer
from restricted support. Hence, it is not possible to rename topics or to
merge them.
}
\examples{
library(stm)
stm <- stm(poliblog5k.docs, poliblog5k.voc, K=25,
           prevalence=~rating, data=poliblog5k.meta,
           max.em.its=2, init.type="Spectral") 
as.LDA(stm, docs = poliblog5k.docs)
}
