% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chainsDistances}
\alias{chainsDistances}
\title{Distances between topic models (chains)}
\usage{
chainsDistances(
  x,
  method = c("euclidean", "hellinger", "cosine", "minMax", "stupidEuclidean",
    "stupidEuclidean2", "stupidEuclideanDistances23JST", "invariantEuclidean")
)
}
\arguments{
\item{x}{a valid \code{multiChains} object, obtained through the estimation of a
topic model using \code{\link[=grow]{grow()}} and the argument \code{nChains} greater than \code{1}.}

\item{method}{the method used to measure the distance between chains.}
}
\value{
a matrix of distance between the elements of \code{w}
}
\description{
Computes the distance between different estimates of a topic
model. Since the estimation of a topic model is random, the results may
largely differ as the process is repeated. This function allows to compute
the distance between distinct realizations of the estimation process.
Estimates are referred to as \emph{chains}.
}
\details{
the \code{method} argument determines how are computed distance. In
general, you would want to use \emph{invariantEuclidean} distance. TODO
}
\examples{
model <- LDA(ECB_speeches)
model <- grow(model, 10, nChains = 5)
chainsDistances(model)

}
\seealso{
\code{\link[=plot.multiChains]{plot.multiChains()}} \code{\link[=chainsScores]{chainsScores()}}
}
\author{
Olivier Delmarcelle
}
