% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{ECB_press_conferences}
\alias{ECB_press_conferences}
\title{Example dataset}
\format{
A \code{quanteda::corpus} object.
}
\source{
\url{www.ecb.europa.eu/press/key/html/index.en.html}.
}
\usage{
data("ECB_press_conferences")
}
\description{
...
}
\examples{
data("ECB_press_conferences")
head(ECB_press_conferences)

}
\keyword{datasets}
