% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{LDA}
\alias{LDA}
\title{Create a sentopic model}
\usage{
LDA(x, K = 5, alpha = 1, beta = 0.01)
}
\arguments{
\item{x}{tokens object containing the texts. A coercion will be attempted if \code{x} is not a tokens.}

\item{K}{the number of topics}

\item{alpha}{the hyperparameter of topic-document distribution}

\item{beta}{the hyperparameter of vocabulary distribution}
}
\value{
An initialized \code{sentopicmodel} object inheriting the relevant model class.
This object corresponds to a Gibbs sampler estimation with zero iterations.
The MCMC can be iterated using the \code{\link[=grow]{grow()}} function.

The estimated distributions of the model are accessible using the dollar-sign operator.
The \code{\link[=topWords]{topWords()}} function easily extract the most probables words of each topic/sentiment.
}
\description{
The set of functions \code{\link[=LDA]{LDA()}}, \code{\link[=JST]{JST()}}, \code{\link[=rJST]{rJST()}} and \code{\link[=sentopicmodel]{sentopicmodel()}} are all wrappers to an unified C++
routine and attempt to replicate their corresponding model.
}
\examples{
data(ECB_speeches)
LDA(ECB_speeches)
rJST(ECB_speeches, lexicon = LoughranMcDonald)
}
\seealso{
Other topic models: 
\code{\link{JST}()},
\code{\link{rJST}()},
\code{\link{sentopicmodel}()}
}
\author{
Olivier Delmarcelle
}
\concept{topic models}
