% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{LoughranMcDonald}
\alias{LoughranMcDonald}
\title{Example lexicon}
\format{
A \code{data.table}.
}
\source{
\url{https://sraf.nd.edu/textual-analysis/resources/}
}
\usage{
data("LoughranMcDonald")
}
\description{
...
}
\examples{
data("LoughranMcDonald")
summary(LoughranMcDonald)

}
\keyword{datasets}
