% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{melt}
\alias{melt}
\title{Temporary melt generic}
\usage{
melt(data, ..., na.rm = FALSE, value.name = "value")
}
\arguments{
\item{data}{...}

\item{...}{...}

\item{na.rm}{...}

\item{value.name}{...}
}
\value{
...
}
\description{
...
}
\author{
Olivier Delmarcelle
}
