% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{melt.sentopicmodel}
\alias{melt.sentopicmodel}
\title{Melt for sentopicmodels}
\usage{
\method{melt}{sentopicmodel}(data, ..., include_docvars = FALSE)
}
\arguments{
\item{data}{...}

\item{...}{...}

\item{include_docvars}{...}
}
\value{
...
}
\description{
...
}
\author{
Olivier Delmarcelle
}
