% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{data}
\name{PicaultRenault_data}
\alias{PicaultRenault_data}
\title{Regression dataset}
\format{
A \code{xts}.
}
\source{
\url{https://sraf.nd.edu/textual-analysis/resources/}
}
\usage{
data("PicaultRenault_data")
}
\description{
...
}
\examples{
data("PicaultRenault_data")
summary(PicaultRenault_data)

}
\keyword{datasets}
