% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rJST}
\alias{rJST}
\title{Create a sentopic model}
\usage{
rJST(
  x,
  lexicon = NULL,
  K = 5,
  S = 3,
  alpha = 1,
  gamma = 5,
  beta = 0.01,
  alphaCycle = 0,
  gammaCycle = 0
)
}
\arguments{
\item{x}{tokens object containing the texts. A coercion will be attempted if \code{x} is not a tokens.}

\item{lexicon}{a \code{quanteda} dictionary with positive and negative categories}

\item{K}{the number of topics}

\item{S}{the number of sentiments}

\item{alpha}{the hyperparameter of topic-document distribution}

\item{gamma}{the hyperparameter of sentiment-document distribution}

\item{beta}{the hyperparameter of vocabulary distribution}

\item{alphaCycle}{integer specifying the cycle size between two updates of the hyperparameter alpha}

\item{gammaCycle}{integer specifying the cycle size between two updates of the hyperparameter alpha}
}
\value{
An initialized \code{sentopicmodel} object inheriting the relevant model class.
This object corresponds to a Gibbs sampler estimation with zero iterations.
The MCMC can be iterated using the \code{\link[=grow]{grow()}} function.

The estimated distributions of the model are accessible using the dollar-sign operator.
The \code{\link[=topWords]{topWords()}} function easily extract the most probables words of each topic/sentiment.
}
\description{
The set of functions \code{\link[=LDA]{LDA()}}, \code{\link[=JST]{JST()}}, \code{\link[=rJST]{rJST()}} and \code{\link[=sentopicmodel]{sentopicmodel()}} are all wrappers to an unified C++
routine and attempt to replicate their corresponding model.
}
\examples{
data(ECB_speeches)
LDA(ECB_speeches)
rJST(ECB_speeches, lexicon = LoughranMcDonald)
}
\seealso{
Other topic models: 
\code{\link{JST}()},
\code{\link{LDA}()},
\code{\link{sentopicmodel}()}
}
\author{
Olivier Delmarcelle
}
\concept{topic models}
