% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{sentiment_series}
\alias{sentiment_series}
\title{Title}
\usage{
sentiment_series(
  x,
  period = c("year", "month", "day"),
  rolling_window = 1,
  scale = TRUE,
  scaling_period = c("1900-01-01", "2099-12-31"),
  as.xts = TRUE,
  ...
)
}
\arguments{
\item{x}{...}

\item{period}{...}

\item{rolling_window}{...}

\item{scale}{...}

\item{scaling_period}{...}

\item{as.xts}{...}

\item{...}{...}
}
\value{
...
}
\description{
Title
}
\seealso{
Other plot functions: 
\code{\link{proportion_topics}()},
\code{\link{sentiment_breakdown}()}
}
