% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentopics.R
\docType{package}
\name{sentopics-package}
\alias{sentopics}
\alias{sentopics-package}
\title{Tools for joining sentiment and topic analysis (sentopics)}
\description{
\strong{sentopics} provides function to easily estimate a range of
topic models and process their output. Particularly, it facilitates the
integration of topic analysis with a time dimension through time-series
generating functions. In addition, \strong{sentopics} interacts with sentiment
analysis to compute the sentiment conveyed by topics. Finally, the package
implements a number of visualization helping interpreting the results of
topic models.
}
\note{
Please cite the package in publications. Use
\code{citation("sentopics")}.
}
\section{Usage}{


Please refer to the vignettes for a comprehensive introduction to the
package functions.
\itemize{
\item \href{../doc/Basic_usage.html}{Basic usage}: Introduction to topic model estimation with \strong{sentopics}
\item \href{../doc/Basic_usage.html}{Topical time series}: Integrate topic analysis with sentiment analysis along a time dimension
}

For further details, you may browse the package \href{../html/00Index.html}{documentation}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odelmarcelle/sentopics}
  \item Report bugs at \url{https://github.com/odelmarcelle/sentopics/issues}
}

}
\author{
\strong{Maintainer}: Olivier Delmarcelle \email{delmarcelle.olivier@gmail.com} (\href{https://orcid.org/0000-0003-4347-070X}{ORCID})

Other contributors:
\itemize{
  \item Samuel Borms \email{samuel.borms@unine.ch} (\href{https://orcid.org/0000-0001-9533-1870}{ORCID}) [contributor]
  \item Chengua Lin \email{chenghua.lin@abdn.ac.uk} (Original JST implementation) [copyright holder]
  \item Yulan He \email{yulan.he@warwick.ac.uk} (Original JST implementation) [copyright holder]
  \item Jose Bernardo (Original JST implementation) [copyright holder]
  \item David Robinson \email{admiral.david@gmail.com} (Implementation of reorder_within()) [copyright holder]
  \item Julia Silge \email{julia.silge@gmail.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID}) (Implementation of reorder_within()) [copyright holder]
}

}
