% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{topWords}
\alias{topWords}
\alias{plot_topWords}
\title{TITLE}
\usage{
topWords(
  x,
  nWords = 10,
  method = c("frequency", "probability", "term-score"),
  output = c("data.frame", "plot", "matrix"),
  subset
)

plot_topWords(
  x,
  nWords = 10,
  method = c("frequency", "probability", "term-score"),
  subset
)
}
\arguments{
\item{x}{a \code{sentopicmodel} created from the \code{\link[=LDA]{LDA()}}, \code{\link[=JST]{JST()}}, \code{\link[=rJST]{rJST()}} or
\code{\link[=sentopicmodel]{sentopicmodel()}} function}

\item{nWords}{the number of top words to extract}

\item{method}{specify if a re-ranking function should be applied before
returning the top words}

\item{output}{determines the output of the function}
}
\value{
the top words of the topic model. Depending on the output chosen, can
result in either a long-style data.frame, a \code{ggplot2} object or a matrix.
}
\description{
Extract the top words in each topic/sentiment from a
\code{sentopicmodel}.
}
\examples{
model <- LDA(ECB_speeches)
model <- grow(model, 10)
topWords(model)
plot_topWords(model)
topWords(model, output = "matrix")
library(ggplot2)
ggplot <- plot_topWords(model)
ggplot \%+\% subset(ggplot$data, L1 \%in\% 1:2)
}
\author{
Olivier Delmarcelle
}
