% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{vocabFromList}
\alias{vocabFromList}
\title{Generate vocabulary from list of words}
\usage{
vocabFromList(list)
}
\arguments{
\item{list}{list of words. First list level determine the topic. Second list
level indicate the sentiment}
}
\value{
an array usable in \code{generateDocuments}
}
\description{
Generate vocabulary from list of words
}
\examples{
sentopics:::vocabFromList(unname(list(
topic1 = list(sent1 = c("tasty", "delicious"), sent2 = c("bad", "smelly")),
topic2 = list(sent1 = c("surprising", "amazing"), sent2 = c("boring", "annoying")),
topic3 = list(sent1 = "fearless", sent2 = "coward")
)))
sentopics:::vocabFromList(lapply(1:3, function(i) c(paste0("topic",i,"w1"))))
}
\keyword{internal}
